Fixed FreeType-DLL loading and used in Gtk4Demo to showcase (#41)
* Simplified names in xml doc comments

* FreeTypeDemo with more fonts (*.ttf and *.otf fonts)

* Fixed null values and nullability in FontFace and ScaledFont

* Added TextExtensions.GetTypedFontFace and used it in FreeTypeDemo

* PlotLightnessCharacteristics with FontFace as parameter

* Native resolver with WindowsAltLibName which is searched first

E.g. for FreeType the DLL in the bin-folder was used, but correct would be the one with lib-prefix when used from msys64-installation. The easiest and best way would be, when the native cairo build with meson would output DLLs with lib-prefix, so these hacks aren't needed at all. But according meson docs that's not possible.

* LightnessPlotWindow with FreeType font (per #ifdef)

* PR feedback