Fixed race condition in DefaultFonts
When multiple threads read a default font property, then the first time the field is null, thus the font is created, and the reference assigned to the field.
In the end, only one instance is referenced by field, all others don't have a root, as here the FontFace is only assigned to the native side via the handle, so nothing keeps a reference to the FontFace, and the object's finalizer will clean it up.

A possible solution would be to keep a reference in the managed side of CairoContext via setting the FontFace. But this is kind of doubling the (basic) logic from the native side to the managed side.
So as solution the static properties are made thread-safe with both execution and publication of the initialization.