UserDataKey fixes
Relevant is only the address of the UserDataKey, so it must be fixed / pinned, and hence can't be a field in a managed object (same goes for a static class), as by a GC the object my be relocated, thus the key gets a new (different) address.

Instead of a GCHandle with pinned target, we allocate the key as native memory.

For the (internal) APIs to make it more obvious that a pinned address needs to be used, changed the argument from `ref UserDataKey` to `UserDataKey*`.