Made the project thread-safe and ready for usage in parallel scenarios like web-servers (#54)
* Added k6 load test for ASP.NET Core demo

* FreeTypeTests added with a single iteration

* More tests to repro the bug

* Fixed FontOptions

* Fixed race condition in DefaultFonts

When multiple threads read a default font property, then the first time the field is null, thus the font is created, and the reference assigned to the field.
In the end, only one instance is referenced by field, all others don't have a root, as here the FontFace is only assigned to the native side via the handle, so nothing keeps a reference to the FontFace, and the object's finalizer will clean it up.

A possible solution would be to keep a reference in the managed side of CairoContext via setting the FontFace. But this is kind of doubling the (basic) logic from the native side to the managed side.
So as solution the static properties are made thread-safe with both execution and publication of the initialization.

* UserDataKey tests

* UserDataKey fixes

Relevant is only the address of the UserDataKey, so it must be fixed / pinned, and hence can't be a field in a managed object (same goes for a static class), as by a GC the object my be relocated, thus the key gets a new (different) address.

Instead of a GCHandle with pinned target, we allocate the key as native memory.

For the (internal) APIs to make it more obvious that a pinned address needs to be used, changed the argument from `ref UserDataKey` to `UserDataKey*`.

* Made FreeType thread safe

* Fixed API of freetype around lock / unlock

* More tests and public property for FreeTypeFont.FreeTypeFontFace

* Use thread local storage for default fonts

* PR feedback