Fixed PdfDocument.RenderToPng
According to https://gitlab.freedesktop.org/cairo/cairo/-/commit/f74b11415a1f7682dd50c222baa8815ef93681dc:
In order to support blend modes, we need to use a clear background whilst rendering the PDF, which inspired the use of DEST_OVER to avoid having to render to an intermediate surface. However, the adjunct of using subpixel text means that we cannot simply using the alpha channel as the sole mask and so need to render to an intermediate and paint it over an opaque background.